주문하기 작성 중

+++ src/main/java/com/lch/topick/web/order/def/repository/OrderListRepository.java
@@ -3,10 +3,13 @@ package com.lch.topick.web.order.def.repository;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import com.lch.topick.web.order.def.entity.OrderList;
 
-public interface OrderListRepository extends JpaRepository<OrderList,Integer> {
+public interface OrderListRepository extends JpaRepository<OrderList,Long> {
 	
-	public List<OrderList> findByMemberId(String memberId);
+	public List<OrderList> findByMemberId( String memberId );
+	
+	
 
 } //interface
